requirements.txt
Django==3.0.8
django-rest-auth==0.9.5
django-settings-export==1.2.1
djangorestframework==3.11.0
psycopg2-binary==2.8.5
python-decouple==3.1
whitenoise==5.1.0
gunicorn==19.9.0
dj-database-url==0.5.0